+++ src/export/export.cpp
@@ -1,0 +1,11 @@
+#include "export.h"
+
+Export::Export() {}
+
+QString Export::getFilter() {
+    return "";
+}
+
+QString Export::fromHtml(QString htmlLine) {
+    return htmlLine;
+}