+++ src/export/exporthtml.cpp
@@ -1,0 +1,11 @@
+#include "exporthtml.h"
+
+ExportHTML::ExportHTML() {}
+
+QString ExportHTML::getFilter() {
+    return QString(tr("HTML pages (%1)")).arg("*.htm *.html");
+}
+
+QString ExportHTML::fromHtml(QString htmlLine) {
+    return htmlLine;
+}