+++ src/export/exportmarkdown.h
@@ -1,0 +1,15 @@
+#ifndef EXPORTMARKDOWN_H
+#define EXPORTMARKDOWN_H
+
+#include "export.h"
+
+class ExportMarkdown : public Export
+{
+public:
+    ExportMarkdown();
+
+    QString getFilter() override;
+    QString fromHtml(QString htmlLine) override;
+};
+
+#endif // EXPORTMARKDOWN_H